refactor(fix): move useHandleRsvScan
